feat: test each jobId in collection and parity with mainnet contracts
