+++ collections.json
@@ -1,23 +1,9 @@
 [
   {
-    "jobIDs": [1, 2, 3, 4, 5, 6, 7, 8, 9],
+    "jobIDs": [1, 2, 3, 5, 6, 7, 9, 10],
     "aggregationMethod": 2,
     "power": 2,
-    "name": "ethCollectionMedian",
-    "tolerance": 500000
-  },
-  {
-    "jobIDs": [10, 11, 12],
-    "aggregationMethod": 2,
-    "power": 2,
-    "name": "RAZORUSD",
-    "tolerance": 500000
-  },
-  {
-    "jobIDs": [13, 14, 15, 16],
-    "aggregationMethod": 2,
-    "power": 2,
-    "name": "UFTUSD",
+    "name": "ETHUSD",
     "tolerance": 500000
   }
 ]
