+++ jobs.json
@@ -59,6 +59,14 @@
     "weight": 1,
     "power": 2,
     "selectorType": 0,
+    "name": "ethusd_ftx",
+    "selector": "result.price",
+    "url": "https://ftx.com/api/markets/ETH/USD"
+  },
+  {
+    "weight": 1,
+    "power": 2,
+    "selectorType": 0,
     "name": "ethusd_bitfinex",
     "selector": "[0]",
     "url": "https://api-pub.bitfinex.com/v2/ticker/tETHUSD"
@@ -70,61 +78,5 @@
     "name": "ethusd_cryptocompare",
     "selector": "[USD]",
     "url": "https://min-api.cryptocompare.com/data/price?fsym=ETH&tsyms=USD"
-  },
-  {
-    "weight": 1,
-    "power": 2,
-    "selectorType": 0,
-    "name": "razorusd_cryptocompare",
-    "selector": "[USD]",
-    "url": "https://min-api.cryptocompare.com/data/price?fsym=RAZOR&tsyms=USD"
-  },
-  {
-    "weight": 1,
-    "power": 2,
-    "selectorType": 0,
-    "name": "razorusd_mexc",
-    "selector": "data[0][last]",
-    "url": "https://www.mexc.com/open/api/v2/market/ticker?symbol=RAZOR_USDT"
-  },
-  {
-    "weight": 1,
-    "power": 2,
-    "selectorType": 0,
-    "name": "razorusd_gateio",
-    "selector": "[0].last",
-    "url": "https://api.gateio.ws/api/v4/spot/tickers?currency_pair=RAZOR_USDT"
-  },
-  {
-    "weight": 1,
-    "power": 2,
-    "selectorType": 0,
-    "name": "uftusd_mexc",
-    "selector": "data[0][last]",
-    "url": "https://www.mexc.com/open/api/v2/market/ticker?symbol=UFT_USDT"
-  },
-  {
-    "weight": 1,
-    "power": 2,
-    "selectorType": 0,
-    "name": "uftusd_gateio",
-    "selector": "[0].last",
-    "url": "https://api.gateio.ws/api/v4/spot/tickers?currency_pair=UFT_USDT"
-  },
-  {
-    "weight": 1,
-    "power": 2,
-    "selectorType": 0,
-    "name": "uftusd_cryptocompare",
-    "selector": "[USD]",
-    "url": "https://min-api.cryptocompare.com/data/price?fsym=UFT&tsyms=USD"
-  },
-  {
-    "weight": 1,
-    "power": 2,
-    "selectorType": 0,
-    "name": "uftusd_binance",
-    "selector": "price",
-    "url": "https://api.binance.com/api/v3/avgPrice?symbol=UFTBUSD"
   }
 ]
